feat: add support for @casl/ability in security validation
If present, req.ability must be an Ability instance and it's used to check if a user can perform the
requested operation, to constraint mongo queries (via CASL conditions), to filter fields in the
responses (via CASL fields)

BREAKING CHANGE: API.securityValidator renamed to API.initSecurity. The old method is still present
but will warn it's deprecated.